POINT: added a cast in dxf_point_write().

+++ src/point.c
@@ -478,7 +478,7 @@ dxf_point_write
         }
         if (point->paperspace == DXF_PAPERSPACE)
         {
-                fprintf (fp->fp, " 67\n%d\n", DXF_PAPERSPACE);
+                fprintf (fp->fp, " 67\n%d\n", (int16_t) DXF_PAPERSPACE);
         }
         fprintf (fp->fp, "  8\n%s\n", point->layer);
         if (strcmp (point->linetype, DXF_DEFAULT_LINETYPE) != 0)
